Remove page name normaliztion
This page name normalization is not correct. We should not assume that all pages with spaces in their name should be renamed. This causes issue between Core and Framework where Framework has no normalization and when it tries to load the list item or verify the page is created, it queries for the page with spaces, but the save method in Core always saves with replacing spaces with dashes.

This also causes issues with links to pages throughout a template that have linked to a page with spaces in it's name.

I believe this normalization was implimented without understanding implications. I feel that if SharePoint allows a user to rename a page to have spaces in the title, then PnP should respect this and not rename the file.